requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain
pydantic
supabase
langgraph>=0.2.26
openai